Added Product Explanation page, documentation, and updated build script to automatically rebuild TypeScript, JavaScript, and SASS files (#27)
# Major Changes
- Added first version of the web site
- Added top navigation bar. The Go to main page button and Sign-in
button do not work but are visible.
- Added Production Explanation Page / Home Page - The page explains what
Goals is and why someone should use it. The Signup button is not hooked
up. You can press it but it does nothing.
		
- Added documentation folder
- Moved the specification HTML and CSS files to the .\Documents folder
- Added a Visio diagram of how the first version of the web site and
backend work.
		
- Updated the web site build script 
- It now can automatically rebuild JavaScript and Sass file changes if
the -BuildAndServeWebSite command line argument is used. The Sass and
JavaScript build windows are in separate terminal tabs.
- Update the build scripts to build the main web site's TypeScript and
Scss files.

# Minor Changes
- Added new Visual Studio Code debug targets (Debug Web Site and Start
web site). Start web site is useful for starting the build script and
the debug web server. In this mode, TypeScript, JavaScript, and SCSS
files are automatically rebuilt when the change. Debug Web Site is
useful for debugging the web site's code when it runs in the browser.

- Update dependencies to latest versions and added React.

Co-authored-by: Ben Ellett <github@benellett.net>